[#31151] fix POST_DYNA_MODA_T backport_v16

+++ bibfor/algorith/statch.F90
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2023 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2024 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -108,8 +108,8 @@ subroutine statch(nbobst, nbpt, temps, dloc, fcho, &
     else
         dt = temps(2)-temps(1)
     end if
-    idebut = int((tdebut-temps(1))/dt)+1
-    ifin = min(int((tfin-temps(1))/dt)+1, nbpt)
+    idebut = nint((tdebut-temps(1))/dt)+1
+    ifin = min(nint((tfin-temps(1))/dt)+1, nbpt)
     nbpas = ifin-idebut+1
     if (nbloc .eq. 0) nbloc = 1
     nbval = nbpas/nbloc
